testing polygammas (63.77% Statements 345/54155.36% Branches 31/5642.86% Functions 3/763.77% Lines 345/541)
